fix(web,docs): relocate fork tests and address review findings
- Move the two workbook Fork tests from the page test to
  workbook-layout-content.test.tsx, where the Fork button now lives. This fixes
  the failing web test suite after the button moved into the layout header.
- PlatformLink sets target="_blank" rel="noopener noreferrer" explicitly so
  every docs-to-platform link (CTA, inline, starter workbook) opens in a new tab
  rather than relying on Fumadocs external-link detection.
- Normalize a trailing slash on PLATFORM_URL so a trailing-slash PLATFORM_BASE_URL
  cannot produce a double slash in generated links.
