+++ apps/web/components/workbook-overview/workbook-layout-content.test.tsx
@@ -1,5 +1,5 @@
 import { AutosaveStatusProvider } from "@/components/shared/autosave/autosave-status-context";
-import { createWorkbook, createWorkbookVersionSummary } from "@/test/factories";
+import { createMarkdownCell, createWorkbook, createWorkbookVersionSummary } from "@/test/factories";
 import { server } from "@/test/msw/server";
 import { render, screen, userEvent, waitFor } from "@/test/test-utils";
 import { describe, it, expect, vi, beforeEach } from "vitest";
@@ -123,5 +123,47 @@ describe("WorkbookLayoutContent", () => {
     // The version cell shows "-" rather than wrongly claiming the workbook is a draft.
     expect(await screen.findByText("-")).toBeInTheDocument();
     expect(screen.queryByText("workbooks.draftVersion")).not.toBeInTheDocument();
+  });
+
+  it("forks the workbook and posts forkedFrom", async () => {
+    const sourceCells = [
+      createMarkdownCell({ id: "source-cell", content: "<p>Source instructions</p>" }),
+    ];
+    const sourceMetadata = { crop: "maize", trialYear: 2026 };
+    const spy = server.mount(contract.workbooks.createWorkbook, {
+      status: 201,
+      body: createWorkbook({ id: "99999999-9999-9999-9999-999999999999" }),
+    });
+    const user = userEvent.setup();
+    renderContent({
+      name: "Distinctive field workbook",
+      description: "A workbook description that must survive forking.",
+      cells: sourceCells,
+      metadata: sourceMetadata,
+    });
+
+    await user.click(await screen.findByRole("button", { name: "workbooks.actions.fork" }));
+
+    await waitFor(() => expect(spy.called).toBe(true));
+    expect(spy.body).toEqual({
+      name: "Fork of Distinctive field workbook",
+      description: "A workbook description that must survive forking.",
+      cells: sourceCells,
+      metadata: sourceMetadata,
+      forkedFrom: "wb-1",
+    });
+  });
+
+  it("shows the fork button to viewers who are not the creator", async () => {
+    vi.mocked(useSession).mockReturnValue({
+      data: { user: { id: "someone-else" } },
+      isPending: false,
+    } as ReturnType<typeof useSession>);
+
+    renderContent({ createdBy: "user-1" });
+
+    expect(
+      await screen.findByRole("button", { name: "workbooks.actions.fork" }),
+    ).toBeInTheDocument();
   });
 });
